feat: SnapShotMap - 경로 표시 이미지 사용 x - kakaoMap 에서 경로를 이미지로 저장 할 수 없음 - 따라서 SnapShotMap 컴포넌트를 활용해서 보여준다.
